move policyItem types to /types to fix circular dep
