+++ src/components/ClaimCard.svelte
@@ -3,9 +3,9 @@ import { getUploadLabel, isEvidenceNeeded } from '../business-rules/claim-payout
 import ClaimBanner from './banners/ClaimBanner.svelte'
 import ClaimCardBanner from './ClaimCardBanner.svelte'
 import { Claim, ClaimItem, ClaimStatus, PayoutOption, ReceiptType } from 'data/claims'
-import type { PolicyItem } from 'data/items'
 import { roleSelection } from 'data/role-policy-selection'
 import { getClaimState, SecondaryClaimStatus, State } from 'data/states'
+import type { PolicyItem } from 'data/types/items'
 import { Card, Button } from '@silintl/ui-components'
 import { createEventDispatcher } from 'svelte'
 import { differenceInSeconds, formatDistanceToNow } from 'date-fns'
