+++ src/components/Datatable/PolicyItems/ItemsTable.svelte
@@ -2,9 +2,9 @@
 import BatchItemClone from './BatchItemClone.svelte'
 import BatchItemDelete from './BatchItemDelete.svelte'
 import CopyTableButton from '../CopyTableButton.svelte'
-import { ClaimItem, incompleteClaimItemStatuses, selectedPolicyClaims } from 'data/claims'
+import { itemIsVehicle } from 'data/items'
 import { getItemState } from 'data/states'
-import { AccountablePerson, editableCoverageStatuses, ItemCoverageStatus, itemIsVehicle, PolicyItem } from 'data/items'
+import { AccountablePerson, editableCoverageStatuses, ItemCoverageStatus, PolicyItem } from 'data/types/items'
 import { isMonthly } from 'helpers/coverage'
 import { formatDate, formatFriendlyDate } from 'helpers/dates'
 import { getItemIcon, hasEnded, willEnd } from './itemTableHelpers'
