+++ src/components/ItemDetails.svelte
@@ -1,8 +1,9 @@
 <script lang="ts">
 import ItemBanner from './banners/ItemBanner.svelte'
 import MessageBanner from './banners/MessageBanner.svelte'
-import { ItemCoverageStatus, PolicyItem, itemIsDenied, itemIsNotInactive } from 'data/items'
+import { itemIsDenied, itemIsNotInactive } from 'data/items'
 import { getPolicyById, loadPolicy, policies, Policy, PolicyType } from 'data/policies'
+import { ItemCoverageStatus, PolicyItem } from 'data/types/items'
 import { getPremiumDescription, getRenewalDate, getStartDate } from 'helpers/coverage'
 import { dateIsInThePast, formatDate } from 'helpers/dates'
 import { formatMoney } from 'helpers/money'
