+++ src/components/banners/ItemBanner.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 import { getItemState, State } from 'data/states'
-import type { ItemCoverageStatus } from 'data/items'
+import type { ItemCoverageStatus } from 'data/types/items'
 import StatusBanner from './StatusBanner.svelte'
 
 export let itemStatus = '' as ItemCoverageStatus
