+++ src/components/forms/items/itemFormHelpers.ts
@@ -1,5 +1,5 @@
 import { categories } from 'data/itemCategories'
-import { ItemCoverageStatus, PolicyItem, NewItemFormData, UpdateItemFormData } from 'data/items'
+import { ItemCoverageStatus, PolicyItem, NewItemFormData, UpdateItemFormData } from 'data/types/items'
 import { assertHas, assertIsFourDigitYear, assertIsLessOrEqual } from '../../../validation/assertions'
 import { get } from 'svelte/store'
 
@@ -10,7 +10,12 @@ export const areMakeAndModelRequired = (item: PolicyItem, categoryId: string): b
   )
 }
 
-export const assembleStatementNameDefault = (make: string, model: string, year: number | undefined, uniqueIdentifier: string) => {
+export const assembleStatementNameDefault = (
+  make: string,
+  model: string,
+  year: number | undefined,
+  uniqueIdentifier: string
+) => {
   const consecutiveSpaces = / {2,}/
   const combinedValues = `${make} ${model} ${year || ''} ${uniqueIdentifier.slice(-6)}`
   return combinedValues.trim().replace(consecutiveSpaces, ' ')
@@ -22,7 +27,11 @@ export const validateForSave = (formData: NewItemFormData | UpdateItemFormData):
   assertHas(formData.name, 'Please specify a statement name')
 }
 
-export const validateForSubmit = (item: PolicyItem, formData: NewItemFormData | UpdateItemFormData, isVehicle: boolean): void => {
+export const validateForSubmit = (
+  item: PolicyItem,
+  formData: NewItemFormData | UpdateItemFormData,
+  isVehicle: boolean
+): void => {
   validateForSave(formData)
   assertIsLessOrEqual(0.01, Number(formData.coverageAmountUSD), 'Please specify the coverage value')
   item.coverage_status !== ItemCoverageStatus.Draft &&
@@ -32,9 +41,6 @@ export const validateForSubmit = (item: PolicyItem, formData: NewItemFormData | 
       'Coverage amount cannot be increased'
     )
   if (isVehicle) {
-    assertIsFourDigitYear(
-      formData.year,
-      "Please enter the vehicle's model year, e.g., 1995 (all four digits)"
-    )
+    assertIsFourDigitYear(formData.year, "Please enter the vehicle's model year, e.g., 1995 (all four digits)")
   }
 }
