+++ src/data/audits.ts
@@ -1,5 +1,5 @@
-import type { PolicyItem } from './items'
 import { CREATE } from 'data'
+import type { PolicyItem } from './types/items'
 import { writable } from 'svelte/store'
 
 export type AuditResult = {
