+++ src/data/claims.ts
@@ -1,8 +1,8 @@
 import type { ClaimData, ClaimItemData } from 'components/forms/claims/types'
 import { CREATE, DELETE, GET, UPDATE } from 'data'
 import type { CoverFile } from 'data/file'
-import type { PolicyItem } from 'data/items'
 import { selectedPolicyId } from 'data/role-policy-selection'
+import type { PolicyItem } from 'data/types/items'
 import { UserAppRole } from 'data/user'
 import { convertToCents } from 'helpers/money'
 import { derived, writable } from 'svelte/store'
