+++ src/pages/admin/home.svelte
@@ -2,9 +2,9 @@
 import { CardsGrid, RecentActivityTable, Row } from 'components'
 import { loading } from 'components/progress'
 import { Claim, statusesAwaitingSignator, statusesAwaitingSteward } from 'data/claims'
-import { ItemCoverageStatus, PolicyItem } from 'data/items'
 import { roleSelection } from 'data/role-policy-selection'
 import { isRecentClaim, isRecentItem, loadRecentActivity, RecentChange, recentChanges } from 'data/recent-activity'
+import { ItemCoverageStatus, PolicyItem } from 'data/types/items'
 import { isUserSteward } from 'data/user'
 import { formatPageTitle } from 'helpers/pageTitle'
 import { customerClaimDetails, itemDetails } from 'helpers/routes'
