+++ src/pages/policies/[policyId]/items/new.svelte
@@ -1,17 +1,10 @@
 <script lang="ts">
 import { Breadcrumb, ItemForm, NoHouseholdIdModal } from 'components'
 import { loadDependents } from 'data/dependents'
-import {
-  addItem,
-  loadItems,
-  NewItemFormData,
-  PolicyItem,
-  UpdateItemFormData,
-  selectedPolicyItems,
-  updateItem,
-} from 'data/items'
+import { addItem, loadItems, selectedPolicyItems, updateItem } from 'data/items'
 import { PolicyType, selectedPolicy, updatePolicy } from 'data/policies'
 import { loadMembersOfPolicy } from 'data/policy-members'
+import type { NewItemFormData, PolicyItem, UpdateItemFormData } from 'data/types/items'
 import { formatPageTitle } from 'helpers/pageTitle'
 import { HOME, items as itemsRoute, itemsCheckout, itemsNew, itemsNewQs } from 'helpers/routes'
 import { goto, metatags, params, redirect } from '@roxi/routify'
